--- a/MaxHeap.h
+++ b/MaxHeap.h
@@ -1,18 +1,17 @@
+#pragma once
 #include <vector>
 #include "PCB.h"
 
-class MaxHeap{
+class MaxHeap {
 private:
     int size;
-
-public:
-    MaxHeap(); 
+public:  
+    MaxHeap();
     vector<PCB*> processList;
 
     int getSize();
     void display();
-
-    void swap(PCB &a, PCB &b);
+    void swap(PCB& a, PCB& b);
     void siftUp(int i);
     void siftDown(int i);
     void insert(PCB& process);