@briandrichey 2021-09-20
On Mon, Sep 20, 2021 at 12:20 PM Brian Richey ***@***.***>
wrote:

>
>
> On Mon, Sep 20, 2021 at 11:34 AM Jake Tremblay ***@***.***>
> wrote:
>
>> 9/20/21
>> ------------------------------
>> You can view, comment on, or merge this pull request online at:
>>
>>   https://github.com/briandrichey/CS433_Op-Sys_Group_/pull/4
>> Commit Summary
>>
>>    - Complete
>>
>> File Changes
>>
>>    - *A* Main.cpp
>>    <https://github.com/briandrichey/CS433_Op-Sys_Group_/pull/4/files#diff-ff9db3cb6025e343dacac7fe1b565394e85d2b614be4ab7b7fedb99d8e8b762d>
>>    (112)
>>    - *M* MaxHeap.cpp
>>    <https://github.com/briandrichey/CS433_Op-Sys_Group_/pull/4/files#diff-3809369f2b71c52ae44e5bdc76a782a98df5cc47621eb9d224381cc0a48d4a2e>
>>    (67)
>>    - *M* MaxHeap.h
>>    <https://github.com/briandrichey/CS433_Op-Sys_Group_/pull/4/files#diff-7a1f13b295c33ffc103e025d35021057e9f3c04288ff77f6b393d9b69016fdc7>
>>    (11)
>>    - *M* PCB.cpp
>>    <https://github.com/briandrichey/CS433_Op-Sys_Group_/pull/4/files#diff-301ac56702ba835da00741456168950831876c1268cb19a2d0397adb63762a2e>
>>    (27)
>>    - *M* PCB.h
>>    <https://github.com/briandrichey/CS433_Op-Sys_Group_/pull/4/files#diff-8f83c16429da37b5cc0546253af1a6f5533fe4984ecbbad7a0dd0ff64af3a6c7>
>>    (15)
>>    - *M* ReadyQueue.cpp
>>    <https://github.com/briandrichey/CS433_Op-Sys_Group_/pull/4/files#diff-4d4cc702fc06feb361e20eec15e143c7916f2c963407986373c95f8dad959da0>
>>    (40)
>>    - *M* ReadyQueue.h
>>    <https://github.com/briandrichey/CS433_Op-Sys_Group_/pull/4/files#diff-528dd334f89a94fe423740666fb311b538a34e57e84a0221198f19239deeb305>
>>    (11)
>>
>> Patch Links:
>>
>>    - https://github.com/briandrichey/CS433_Op-Sys_Group_/pull/4.patch
>>    - https://github.com/briandrichey/CS433_Op-Sys_Group_/pull/4.diff
>>
>> —
>> You are receiving this because you are subscribed to this thread.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/briandrichey/CS433_Op-Sys_Group_/pull/4>, or
>> unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/AO3NEXNZB3PBECVHSAJS33DUC55FFANCNFSM5EMRYVAA>
>> .
>>
>

#include "ReadyQueue.h"
#include <time.h>

//First Test: Tests the creation and removal of PCB into/out of the ReadyQueue. Also uses the display function to output the entire contents of the ReadyQueue
void test1() {
	ReadyQueue q1;

	PCB testPCB(15, 15, NEW);
	PCB testPCB2(6, 6, NEW);
	PCB testPCB3(23, 23, NEW);
	PCB testPCB4(39, 39, NEW);
	PCB testPCB5(8, 8, NEW);


	q1.addPCB(testPCB);
	q1.addPCB(testPCB2);
	q1.addPCB(testPCB3);
	q1.addPCB(testPCB4);
	q1.addPCB(testPCB5);
	q1.removePCB();
	q1.display();


	PCB testPCB6(47, 47, NEW);
	PCB testPCB7(1, 1, NEW);
	PCB testPCB8(37, 37, NEW);
	PCB testPCB9(5, 5, NEW);

	q1.addPCB(testPCB6);
	q1.addPCB(testPCB7);
	q1.addPCB(testPCB8);
	q1.addPCB(testPCB9);
	q1.removePCB();
	q1.display();

	PCB testPCB10(43, 43, NEW);
	PCB testPCB11(17, 17, NEW);
	PCB testPCB12(32, 32, NEW);
	PCB testPCB13(12, 12, NEW);
	PCB testPCB14(19, 19, NEW);

	q1.addPCB(testPCB10);
	q1.addPCB(testPCB11);
	q1.addPCB(testPCB12);
	q1.addPCB(testPCB13);
	q1.addPCB(testPCB14);

	while (q1.size() != 0) {
		q1.removePCB();
	}
}

//Second Test: Tests the execution speed of the program. We use a coin flip type method for determining if we should try to add or remove from queue.
void test2() {
	ReadyQueue q2;

	for (int i = 0; i < 100; i++) {
		int randomPrio = rand() % 50;
		PCB tempPCB(i, randomPrio, NEW);
		q2.addPCB(tempPCB);
	}

	int uniqueID = 100; 
	for (int i = 0; i < 1000000; i++) {

		int coinFlip = rand() % 2 + 1;


		if (coinFlip == 1) { //add a pcb
			int randomPrio = rand() % 50;
			PCB tempPCB(++uniqueID, randomPrio, NEW);
			q2.addPCB(tempPCB);
		}
		else if (coinFlip == 2) { //remove a pcb
			q2.removePCB();
		}
	}
}

//used this resource ->(https://www.geeksforgeeks.org/measure-execution-time-with-high-precision-in-c-c/)<- to help with timing the tests
//main will just call the test1() and test2() to make main a little cleaner
int main() {

	cout << "/////Project Members: Jake Tremblay, Brian Richey, Danny Ha/////" << endl;
	cout << "/////Project emulates processes being entered and run from a priority queue/////" << endl;

	//-------TEST 1---------
	clock_t start, end;
	start = clock();

	test1();

	end = clock();

	double elapsedTime = double(end - start) / double(CLOCKS_PER_SEC);
	cout << "Time taken in test 1: " << fixed << elapsedTime;
	cout << " sec " << endl;

	//-------TEST 2---------
	clock_t start2, end2;
	start2 = clock();

	test2();

	end2 = clock();

	double elapsedTime2 = double(end2 - start2) / double(CLOCKS_PER_SEC);
	cout << "Time taken in test 2: " << fixed << elapsedTime2;
	cout << " sec " << endl;

	return 0;
} 
#include "MaxHeap.h"
#include "ReadyQueue.h"


MaxHeap::MaxHeap(){
//Default constructor, sets size of heap to 0
MaxHeap::MaxHeap() {
    size = 0;
}

int MaxHeap::getSize(){
//Returns size of heap
int MaxHeap::getSize() {
    return size;
}

bool MaxHeap::isLeaf(int i){
    return (i < size/2) ? false : true; 
//Checks if the current index i is the index of a leaf node in our heap, if an index i is equal to half the size of the heap, that node is a leaf
bool MaxHeap::isLeaf(int i) {
    return (i < size / 2) ? false : true;
}

void MaxHeap::siftDown(int i){
    while(!isLeaf(i)){
//Sift down uses the root (or in our case just index 0) and will test it against its children to see if it needs to go sift 'downwards' in the heap
void MaxHeap::siftDown(int i) {
    while (!isLeaf(i)) {
        int leftIndex = (2 * i) + 1;
        int rightIndex = (2 * i) + 2;
        int larger = leftIndex;

        if(rightIndex < size){
            larger = (processList[leftIndex] < processList[rightIndex]) ? rightIndex : leftIndex;
        if (rightIndex < size) {
            larger = (processList[leftIndex]->getPrio() < processList[rightIndex]->getPrio()) ? rightIndex : leftIndex;
        }
        if(processList[i] >= processList[larger]){
        if (processList[i]->getPrio() >= processList[larger]->getPrio()) {
            break;
        }

        swap(*processList[i], *processList[larger]);
        i = larger;
    }
}

void MaxHeap::siftUp(int i){
//Sift up will normally use index i = size and sift the node up the heap to proper position, similar to siftDown
void MaxHeap::siftUp(int i) {
    int parentIndex = (i - 1) / 2;

    while(i > 0 && processList[parentIndex]->getPrio() > processList[i]->getPrio()){
    while (i > 0 && processList[parentIndex]->getPrio() < processList[i]->getPrio()) {
        swap(*processList[i], *processList[parentIndex]);
        i = parentIndex;
        parentIndex = (i - 1) / 2;
    }
}

void MaxHeap::display(){
    //remember we need to print out a header with our info on it
//Display all elements currently within the readyQueue
void MaxHeap::display() {
    cout << "### Displaying ReadyQueue ###" << endl;
    for (int i = 0; i < processList.size() && processList[i] != NULL; i++) {
        processList[i]->display();
    }
}

PCB* MaxHeap::getMax(){
    return processList[0];
//getMax returns a PCB pointer that is the highest priority process in the heap, or just the first element
PCB* MaxHeap::getMax() {
    if (size <= 0) {
        return NULL;
    }

    PCB* target = processList[0];
    target->setState(RUNNING);

    processList[0] = processList[--size];
    processList.pop_back();
    siftDown(0);

    return target;
}

void MaxHeap::swap(PCB& a, PCB& b){
//simply swap the address of PCB a and PCB b
void MaxHeap::swap(PCB& a, PCB& b) {
    PCB temp = b;
    b = a;
    a = temp;
}

void MaxHeap::insert(PCB& process){
    //need to follow older code from cs311 - asgn 4
}
//inserts PCB address into the processList vector<PCB&> and invokes the siftUp on i = size
void MaxHeap::insert(PCB& process) {
    processList.push_back(&process);
    int i = size;
    siftUp(i);
    size++;
}  
#pragma once
#include <vector>
#include "PCB.h"

class MaxHeap{
class MaxHeap {
private:
    int size;

public:
    MaxHeap(); 
public:  
    MaxHeap();
    vector<PCB*> processList;

    int getSize();
    void display();

    void swap(PCB &a, PCB &b);
    void swap(PCB& a, PCB& b);
    void siftUp(int i);
    void siftDown(int i);
    void insert(PCB& process);
#pragma once
#include <iostream>

using namespace std;

enum processStates{
enum processStates {
    //different process states for the PCB
    NEW,
    READY,
@@ -11,16 +12,16 @@ enum processStates{
    TERMINATED
};

class PCB{
class PCB {
    int id;
    int prio;
    processStates state;

    public:
public:
    PCB();
    PCB(int id, int prio, processStates state);
    ~PCB();
    

    void setID(int id);
    int getID();

@@ -29,6 +30,6 @@ class PCB{

    void setState(processStates state);
    processStates getState();
    
    ostream& operator << (ostream&,const PCB&);
};

    void display();
}; 
#include "PCB.h"

PCB::~PCB(){    
//Default deconstructor for PCB objects
PCB::~PCB() {

}

PCB::PCB(){
//Default construtor for PCB
PCB::PCB() {
    id = 0;
    prio = 0;
    state = NEW;
}

PCB::PCB(int id, int prio, processStates state){
//Parameterized constructor for PCB objects
PCB::PCB(int id, int prio, processStates state) {
    this->id = id;
    this->prio = prio;
    this->state = state;
}

void PCB::setID(int id){
//PCB 'getters and setters' to facilitate creation/update of PCB
void PCB::setID(int id) {
    this->id = id;
}

void PCB::setPrio(int prio){
void PCB::setPrio(int prio) {
    this->prio = prio;
}

void PCB::setState(processStates state){
void PCB::setState(processStates state) {
    this->state = state;
}

int PCB::getID(){
int PCB::getID() {
    return this->id;
}

int PCB::getPrio(){
int PCB::getPrio() {
    return this->prio;
}

processStates PCB::getState(){
processStates PCB::getState() {
    return this->state;
}


void PCB::display() {
    cout << "PCB ID:" << this->getID() << " /// " << "Priority: " << this->getPrio() << " /// " << "State: " << this->getState() << endl;
} 
#include "ReadyQueue.h"

ReadyQueue::~ReadyQueue(){
    //delete heap here
}
//Default deconstructor
ReadyQueue::~ReadyQueue() {  }

ReadyQueue::ReadyQueue(){
    //not sure what to put in ready queue constructor?
    //vector is made in the maxheap.cpp and maxheap is created readyqueue.h
//Creating the ReadyQueue will invoke creation of a new MaxHeap
ReadyQueue::ReadyQueue() {
    processMaxHeap = new MaxHeap();
}

void ReadyQueue::addPCB(PCB& process){
    //adding to queue -> stored as heap

//Takes a PCB address as parameter, sets processState to READY, and invokes the insert() function from MaxHeap
void ReadyQueue::addPCB(PCB& process) {
    process.setState(READY);
    processMaxHeap->insert(process);

    int numID = process.getID();
    cout<<"PCB "<< numID <<" added to priority queue."<<endl;

    processMaxHeap->insert(process);
}

//should just be able to call maxheap.extractMax, set state to running,
PCB* ReadyQueue::removePCB(){

    cout<<"Highest priority PCB removed from priority queue."<<endl;

//Returns a pointer to PCB, invokes getMax() function from MaxHeap
PCB* ReadyQueue::removePCB() {
    return processMaxHeap->getMax();
}

int ReadyQueue::size(){
    return processMaxHeap->getSize(); //return how many process in the queue
//Simply calls MaxHeap getSize() function and returns it
int ReadyQueue::size() {
    return processMaxHeap->getSize();
}

void ReadyQueue::display(){
    //print the contents of the heap
//Displaying the whole PCB ReadyQueue will display all READY processes
void ReadyQueue::display() {
    processMaxHeap->display();
}

} 
/**
 *  The ready queue will implement a max heap, so we need all the heap functions 
*/

#pragma once
#include "PCB.h"
#include "MaxHeap.h"

class ReadyQueue{
class ReadyQueue {

public:
    ReadyQueue();
    ~ReadyQueue();  
    ~ReadyQueue();

    MaxHeap * processMaxHeap; //heap is going to be used in ReadyQueue to enable us to easily extract highest prio
    MaxHeap* processMaxHeap; //heap is going to be used in ReadyQueue to enable us to easily extract highest prio

    void addPCB(PCB& process);
    PCB* removePCB(); //should remove highest prio, or top of heap